requirements.txt
jupyter-book
matplotlib
numpy
ghp-import
pimoroni-sgp30
sparkfun-qwiic-bme280
click
gpiozero
dash
pandas

